chrome/content/mochitest/tests/test_elemhide.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Element hiding tests</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>        

  <script type="application/x-javascript;version=1.7" src="../httpd.js"></script>
  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="visibility: hidden;">
    <iframe id="frame" onload="runNextTest()"></iframe>
  </div>

  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    prepareFilterComponents();
    preparePrefs();

    let tests = [
      [[], "visible visible"],
      [["#div(test1)"], "hidden visible"],
      [["localhost#div(test1)"], "hidden visible"],
      [["localhost#div(test1)", "foo,foo2#p(test2)"], "hidden visible"],
      [["localhost,foo#div(test1)", "foo,localhost#p(test2)"], "hidden hidden"],
      [["localhost#div(test1)", "localhost#p(test2)"], "hidden hidden"],
      [["foo#div(test1)", "foo#p(test2)"], "visible visible"],

      [["localhost#div(testClass)"], "hidden visible"],
      [["localhost#p(testClass)"], "visible hidden"],
      [["localhost#*(testClass)"], "hidden hidden"],
      [["localhost#div(testClass)", "localhost#p(test2)"], "hidden hidden"],
      [["localhost#p(testClass)", "localhost#p(test2)"], "visible hidden"],
      [["localhost#p(testClass)(test2)"], "visible visible"],   // this filter is invalid, must be ignored

      [["localhost#*(id^=test)"], "hidden hidden"],
      [["localhost#p(id^=test)"], "visible hidden"],
      [["localhost#*(id$=2)"], "visible hidden"],
      [["localhost#p(id$=2)"], "visible hidden"],
      [["localhost#div(id$=2)"], "visible visible"],

      [["localhost#*(test1)(id^=test)"], "hidden visible"],
      [["localhost#*(testClass)(id^=test)"], "hidden hidden"],
      [["localhost#p(testClass)(id^=test)"], "visible hidden"],
      [["localhost#*(test1)(id$=2)"], "visible visible"],
      [["localhost#*(testClass)(id$=2)"], "visible hidden"],
      [["localhost#p(testClass)(id$=2)"], "visible hidden"],

      [["localhost#*(test1)(id^=test)(id$=2)"], "visible visible"],
      [["localhost#*(test1)(id^=test)(id$=1)"], "hidden visible"],
      [["localhost#p(test1)(id^=test)(id$=1)"], "visible visible"],
      [["localhost#div(test1)(id^=test)(id$=1)"], "hidden visible"],
      [["localhost#*(id^=test)(id$=2)"], "visible hidden"],
      [["localhost#*(id^=test)(id$=1)"], "hidden visible"],
      [["localhost#p(id^=test)(id$=1)"], "visible visible"],
      [["localhost#div(id^=test)(id$=1)"], "hidden visible"],

      [["localhost##div#test1"], "hidden visible"],
      [["localhost##p.testClass"], "visible hidden"],
      [["localhost##div#test1, p.testClass"], "hidden hidden"],
      [["localhost##div#test1", "localhost##p.testClass"], "hidden hidden"],
      [["localhost##.testClass"], "hidden hidden"],

      [["~localhost##div#test1"], "visible visible"],
      [["foo,~localhost##div#test1"], "visible visible"],
      [["localhost,~foo##div#test1"], "hidden visible"],
    ];
    let currentTest = -1;
    let currentRun = -1;

    function runNextTest()
    {
      if (currentRun < 0)
        return;

      if (currentTest >= tests.length - 1)
      {
        if (++currentRun <= 5)
        {
          currentTest = -1;
          runNextTest();
        }
        else
          SimpleTest.finish();
        return;
      }

      currentTest++;
      ElemHide.clear();
      whitelistMatcher.clear();

      let [filters, expected] = tests[currentTest];
      for each (let filter in filters)
        ElemHide.add(Filter.fromText(filter));
      ElemHide.apply();
      if (currentRun == 2)
        whitelistMatcher.add(Filter.fromText("@@|chrome://mochikit/$document"));
      else if (currentRun == 3)
        whitelistMatcher.add(Filter.fromText("@@|chrome://mochikit/$~document"));
      else if (currentRun == 4)
        whitelistMatcher.add(Filter.fromText("@@|chrome://mochikit/$elemhide"));
      else if (currentRun == 5)
        whitelistMatcher.add(Filter.fromText("@@||localhost^$elemhide"));

      // Reset the loaded document
      let body = document.getElementById("frame").contentDocument.body;
      body.innerHTML = body.innerHTML;

      if (currentRun == 2 || currentRun == 5)
        expected = "visible visible";   // Second and fifth runs are whitelisted, nothing should be hidden

      body.offsetHeight;    // force reflow
      setTimeout(function()
      {
        let doc = document.getElementById("frame").contentDocument;
        let result = (doc.getElementById("test1").offsetHeight > 0 ? "visible" : "hidden") + " " + (doc.getElementById("test2").offsetHeight > 0 ? "visible" : "hidden");
        is(result, expected, "Run " + currentRun + "\n" + filters.join("\n"));
        runNextTest();
      }, 0);
    }

    function start()
    {
      window.addEventListener("unload", stop, false);
      server.start(1234);

      server.registerPathHandler("/test", function(metadata, response)
      {
        let body = '<div id="test1" class="testClass">foo</div><p id="test2" class="testClass">bar</p>';
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");
        response.bodyOutputStream.write(body, body.length);
      });

      currentRun = 1;

      let frame = document.getElementById("frame");
      frame.contentWindow.location.href = "http://localhost:1234/test";
    }

    function stop()
    {
      server.stop();
    }

    let server = new nsHttpServer();
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(start);
  </script>
  </pre>
</body>
</html>
